Initial project content
Using the smallstep/workflows and smallstep/cli projects
for inspiration, this commit contains the "base" files
intended to be importable by other projects that I maintain.

The goal is to centralize the bulk of the logic currently
replicated (and duplicated) across many different projects.

The primary goal is to reduce the time required to deploy new CI
behavior across all projects.