Explicitly use specific release version
Swap from hard-coded version (used as a workaround) of images
from the atc0005/go-ci project to using a GitHub Actions Workflow
variable set at the top of each workflow.

As of this commit we've switched back to the `latest` tag for
each image, but we can now easily switch all atc0005/go-ci images
to a specific version for testing purposes in the future.

refs GH-56
refs https://github.com/atc0005/go-ci/issues/848