Add job to evaluate Go 1.22 loop behavior changes
Use Go 1.21 series "mirror" image to perform builds using the
`loopvar` compiler flag. This new workflow job will be ued to
identify loops that compile differently using the new loop
semantics introduced as a preview in Go 1.21 and enabled by
default with the release of Go 1.22.

This new workflow job is set to `continue-on-error` in order to
verify that it works as intended. Once it has been proven to
work as intended `continue-on-error` will be set to `false`
and any identified loop behavior changes will be considered
a real CI failure requiring resolution.

refs GH-145