Reuse tag instead of calculating for release build
The Makefile used by dependenent projects attempts to reuse the
`REPO_VERSION` environment variable if set, otherwise falls back
to generating a new value using `git-describe-semver`.

This tool becomes confused if two tags point to the same commit
and does not know which tag to list (ideally the latest stable
tag if matching the same commit as a RC tag) so we work around
that by always having release builds use the tag provided by
GitHub Actions (which matches exactly the tag pushed).